==> scripts/CU_mosaic/00_Data_prep_mosaic_use_case.ipynb <==
# Prepare Mosaic use case from NeurIPS competition data
Author: Pia Rautenstrauch

Date: 4th of February, 2023

- Rationale: use only a single paired data set, three unpaired data sets, as paired assays continue to be more expensive and technically challenging, so it is more likely we only have a few paired data sets
- Donor 1 measured at s1, s2, s4

## Control
- s1: atac only 
- s2: gex only
- s3: atac only
- s4: paired

## A
- s1: atac only - subsampled atac
- s2: gex only
- s3: atac only - subsampled atac 
- s4: paired

## B
- s1: atac only 
- s2: gex only
- s3 atac only 
- s4: paired - subsampled atac 

# Imports
import liam_NeurIPS2021_challenge_reproducibility
import scanpy as sc
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
import torch
import anndata as ad
import os
import scvi

# Load data
ad_mod1 = ad.read_h5ad("./../../data/original/neurips_competition/openproblems_bmmc_multiome_phase2/openproblems_bmmc_multiome_phase2.censor_dataset.output_mod1.h5ad")
ad_mod2 = ad.read_h5ad("./../../data/original/neurips_competition/openproblems_bmmc_multiome_phase2/openproblems_bmmc_multiome_phase2.censor_dataset.output_mod2.h5ad")

ad_mod1.obs["sample"] = ad_mod1.obs["batch"]
ad_mod1.obs["donor"] = ad_mod1.obs["batch"].apply(lambda x: x.split("d")[1])
ad_mod1.obs["site"] = ad_mod1.obs["batch"].apply(lambda x: x.split("d")[0])

ad_mod2.obs["sample"] = ad_mod2.obs["batch"]
ad_mod2.obs["donor"] = ad_mod2.obs["batch"].apply(lambda x: x.split("d")[1])
ad_mod2.obs["site"] = ad_mod2.obs["batch"].apply(lambda x: x.split("d")[0])

ad_mod1.obs['site'].value_counts()

# Control 

ad_mod1.obs['mod'] = 'paired'

ad_mod1.write_h5ad("./../../data/derived/neurips_competition/mosaic_use_case_full.output_mod1.h5ad")

ad_mod2.write_h5ad("./../../data/derived/neurips_competition/mosaic_use_case_full.output_mod2.h5ad")

# sites with d1 measured: s1, s2, s4

# s1 and s3: only atac

ad_mod1.obs['mod'] = ad_mod1.obs['site'].apply(lambda x: 'atac' if ((x == 's1') | (x == 's3')) else 'gex' if x == 's2' else 'paired')

ad_mod1.layers["counts"][(ad_mod1.obs['site'] == 's1') | (ad_mod1.obs['site'] == 's3')] = 0

ad_mod1.layers["counts"][(ad_mod1.obs['site'] == 's1') | (ad_mod1.obs['site'] == 's3')].max()

ad_mod1.layers["counts"].max()

# s2: only gex

ad_mod2.X[ad_mod2.obs['site'] == 's2'] = 0

ad_mod2.X[ad_mod2.obs['site'] == 's2'].max()

ad_mod2.X.max()

ad_mod1.obs['site'].value_counts()

ad_mod1.obs['mod'].value_counts()

ad_mod1.write_h5ad("./../../data/derived/neurips_competition/mosaic_use_case_a.output_mod1.h5ad")

ad_mod2.write_h5ad("./../../data/derived/neurips_competition/mosaic_use_case_a.output_mod2.h5ad")

ad_mod1.write_h5ad("./../../data/derived/neurips_competition/mosaic_use_case_b.output_mod1.h5ad")

ad_mod2.write_h5ad("./../../data/derived/neurips_competition/mosaic_use_case_b.output_mod2.h5ad")

# Downsample ATAC modality by dropping non-zero (binary) observations 

Aim: Investigate how a low quality/coverage modality impacts joint vs. concatenated models

This is a heuristic strategy mimicing lower coverage data. If one had indeed measured a lower coverage dataset, the features selected would differ due to differences in the signal to noise ratios. Here, we called features with full "high quality" dataset, and drop observations from that.

Randomly drop
- 90% 

of observations.

# Create multiple subsamples for 10% 

def which_observations_to_drop(indices, percent):
    """
    Helper function to drop certain percent of non-zero observations from a sparse array.
    
    Parameters
    ----------
    indices: list
        (sorted) list of column indices of non-zero elements (per cell)
    
    percent: float
        How many nonzero observations to drop (percent).
    
    Returns
    -------
    np.array
        Sorted array of nonzero indices to drop.
    """
    # Convert list to array
    arr = np.array(indices)
    # How many nonzero observations to drop - absolute
    drop = int(len(indices) * percent)
    # Randomly permute the observations
    np.random.shuffle(arr)
    # Return sorted array of indices of non-zero elements to drop
    return np.sort(arr[:drop])
    

seeds = [8831, 234, 11, 9631, 94]

# a
# Set random seed for reproducibility
full = ad.read_h5ad("./../../data/derived/neurips_competition/mosaic_use_case_a.output_mod2.h5ad")
full.obs['num_index'] = np.arange(0,full.shape[0])

nonzero_index_list_per_cell = full.X.tolil().rows
# List of Lists Format (LIL) format
# each row is a Python list (sorted) of column indices of non-zero elements
# https://scipy-lectures.org/advanced/scipy_sparse/lil_matrix.html?highlight=lil

input_backup = full.copy()

## Subsampling to 10% of the data
### In other words dropping 90% of the observations
for seed in seeds:
    print(seed)
    np.random.seed(seed)

    full = input_backup.copy()

    # Iterate over cells from site 1 and site 3
    for cell in full.obs[(full.obs['site'] == 's1') | (full.obs['site'] == 's3')]['num_index'].values.tolist():
        drop = which_observations_to_drop(nonzero_index_list_per_cell[cell], 0.9)
        # Which elements to set to zero (inplace)
        full.X[cell, drop.tolist()] = 0

    full.write_h5ad("./../../data/derived/neurips_competition/mosaic_use_case_a.output_mod2_10_subsample_seed_{}.h5ad".format(seed))


    sum_after = full.X.sum(axis=1)

    sum_before = input_backup.X.sum(axis=1)

    print(sum_before)
    print(sum_after)
    print()


# b
# Set random seed for reproducibility
full = ad.read_h5ad("./../../data/derived/neurips_competition/mosaic_use_case_b.output_mod2.h5ad")
full.obs['num_index'] = np.arange(0,full.shape[0])

nonzero_index_list_per_cell = full.X.tolil().rows
# List of Lists Format (LIL) format
# each row is a Python list (sorted) of column indices of non-zero elements
# https://scipy-lectures.org/advanced/scipy_sparse/lil_matrix.html?highlight=lil

input_backup = full.copy()

## Subsampling to 10% of the data
### In other words dropping 90% of the observations
for seed in seeds:
    print(seed)
    np.random.seed(seed)

    full = input_backup.copy()

    # Iterate over cells from site 4
    for cell in full.obs[full.obs['site'] == 's4']['num_index'].values.tolist():
        drop = which_observations_to_drop(nonzero_index_list_per_cell[cell], 0.9)
        # Which elements to set to zero (inplace)
        full.X[cell, drop.tolist()] = 0

    full.write_h5ad("./../../data/derived/neurips_competition/mosaic_use_case_b.output_mod2_10_subsample_seed_{}.h5ad".format(seed))


    sum_after = full.X.sum(axis=1)

    sum_before = input_backup.X.sum(axis=1)

    print(sum_before)
    print(sum_after)
    print()
